Move test from test_multiclass to test_base
