+++ sklearn/tests/test_multiclass.py
@@ -29,7 +29,6 @@ from sklearn.multiclass import (
 from sklearn.naive_bayes import MultinomialNB
 from sklearn.neighbors import KNeighborsClassifier
 from sklearn.pipeline import Pipeline, make_pipeline
-from sklearn.preprocessing import LabelEncoder, StandardScaler
 from sklearn.svm import SVC, LinearSVC
 from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
 from sklearn.utils import (
@@ -81,22 +80,6 @@ def test_check_classification_targets():
     msg = type_of_target(y)
     with pytest.raises(ValueError, match=msg):
         check_classification_targets(y)
-
-
-def test_conditional_attrs_not_in_dir():
-    # Test that __dir__ includes only relevant attributes. #28558
-
-    encoder = LabelEncoder()
-    assert "set_output" not in dir(encoder)
-
-    scalar = StandardScaler()
-    assert "set_output" in dir(scalar)
-
-    svc = SVC(probability=False)
-    assert "predict_proba" not in dir(svc)
-
-    svc.probability = True
-    assert "predict_proba" in dir(svc)
 
 
 def test_ovr_ties():
